lesson-05-api-integration/exercises/starter/requirements.txt
google-adk>=1.17.0
google-cloud-secret-manager
requests
